fibonacci function second fixed

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,16 +6,16 @@
 int main(void)
 {
 	int c;
-	int n1, n2, fn;
+	long int n1, n2, fn;
 
 	n1 = 1;
 	n2 = 2;
 
-	printf("%d, %d", n1, n2);
+	printf("%ld, %ld", n1, n2);
 	for (c = 0; c < 48; c++)
 	{
 		fn = n1 + n2;
-		printf(", %d", fn);
+		printf(", %ld", fn);
 		n1 = n2;
 		n2 = fn;
 	}
